+++ packages/rangelink-vscode-extension/src/types/index.ts
@@ -1,1 +1,3 @@
-export { RangeLinkTerminalLink } from './RangeLinkTerminalLink';
+export type { RangeLinkClickArgs } from './RangeLinkClickArgs';
+export type { RangeLinkTerminalLink } from './RangeLinkTerminalLink';
+export type { WithRequired } from './WithRequired';